fixes #2247 create router-config module to share with light-lambda-native
