+++ router-config/src/main/resources/config/oauthServer.yml
