+++ router-config/src/main/resources/config/pathService.yml
