+++ router-config/src/main/resources/config/serviceDict.yml
