+++ router-config/src/test/resources/config/router.yml
@@ -1,0 +1,80 @@
+---
+# Light Router Configuration
+# As this router is built to support discovery and security for light-4j services,
+# the outbound connection is always HTTP 2.0 with TLS enabled.
+# If HTTP 2.0 protocol will be accepted from incoming request.
+http2Enabled: false
+
+# If TLS is enabled when accepting incoming request. Should be true on test and prod.
+httpsEnabled: true
+
+# Max request time in milliseconds before timeout to the server.
+maxRequestTime: 1000
+
+# If a particular downstream service has different timeout than the above global definition, you can
+# add the path prefix and give it another timeout in millisecond. For downstream APIs not defined here,
+# they will use the global timeout defined in router.maxRequestTime.
+pathPrefixMaxRequestTime: ${router.pathPrefixMaxRequestTime:}
+
+# Connections per thread.
+connectionsPerThread: 10
+
+# The max queue size for the requests if there is no connection to the downstream API in the connection pool.
+# The default value is 0 that means there is queued requests. As we have maxConnectionRetries, there is no
+# need to use the request queue to increase the memory usage. It should only be used when you see 503 errors
+# in the log after maxConnectionRetries to accommodate slow backend API.
+maxQueueSize: ${router.maxQueueSize:0}
+
+# Soft max connections per thread.
+softMaxConnectionsPerThread: 5
+
+# Rewrite Host Header with the target host and port and write X_FORWARDED_HOST with original host
+rewriteHostHeader: true
+
+# Reuse XForwarded for the target XForwarded header
+reuseXForwarded: false
+
+# Max Connection Retries
+maxConnectionRetries: 3
+
+# allowed host list. Use Regex to do wildcard match
+hostWhitelist: ${router.hostWhitelist:}
+
+# support serviceId in the query parameter for routing to overwrite serviceId in header routing.
+# by default, it is false and should not be used unless you are dealing with a legacy client that
+# does not support header manipulation. Once this flag is true, we are going to overwrite the header
+# service_id derived with other handlers from prefix, path, endpoint etc.
+serviceIdQueryParameter: ${router.serviceIdQueryParameter:false}
+
+# URL rewrite rules, each line will have two parts: the regex patten and replace string separated
+# with a space. The light-router has service discovery for host routing, so whe working on the
+# url rewrite rules, we only need to create about the path in the URL.
+# Test your rules at https://www.freeformatter.com/java-regex-tester.html
+urlRewriteRules: ${router.urlRewriteRules:}
+
+# Method rewrite rules for legacy clients that do not support DELETE, PUT, and PATCH HTTP methods to
+# send a request with GET and POST instead. The gateway will rewrite the method from GET to DELETE or
+# from POST to PUT or PATCH. This will be set up at the endpoint level to limit the application.
+# The format of the rule will be "endpoint-pattern source-method target-method". Please refer to test
+# values.yml for examples. The endpoint-pattern is a pattern in OpenAPI specification. The following is
+# a list of examples:
+# /v1/pets/{petId} PATCH POST
+# /v1/pets PUT POST
+# Note: you cannot rewrite a method with a body to a method without a body or vice versa.
+methodRewriteRules: ${router.methodRewriteRules:}
+
+# Query parameter rewrite rules for client applications that send different query parameter keys or values
+# than the target server expecting. When overwriting a value, the key must be specified in order to
+# identify the right query parameter. If only the ok and nk are specified, the router will rewrite the
+# query parameter key ok with different key nk and keep the value.
+# The format of the rule will be a map with the path as the key. Please refer to test values.yml for
+# examples.
+queryParamRewriteRules: ${router.queryParamRewriteRules:}
+
+# Header rewrite rules for client applications that send different header keys or values than the target
+# server expecting. When overwriting a value, the key must be specified in order to identify the right
+# header. If only the ok and nk are specified, the router will rewrite the header key ok with different
+# key nk and keep the value.
+# The format of the rule will be a map with the path as the key. Please refer to test values.yml for
+# examples.
+headerRewriteRules: ${router.headerRewriteRules:}
